Tests/LibWeb/Text/expected/Worker/Worker-script-fetch-referrer.txt

Got referrer: http://127.0.0.1/updated-via-history
